feat(readline): implement zn_readline function for reading lines from socket
