+++ libzappy_net/Makefile
@@ -41,7 +41,8 @@ SRC_C = $(SRC_DIR)/libzappy_net.c \
         $(SRC_DIR)/zappy_net_ringbuf_read.c \
         $(SRC_DIR)/zappy_net_ringbuf_flush.c \
         $(SRC_DIR)/zappy_net_ringbuf_read_fd.c \
-        $(SRC_DIR)/zappy_net_io.c
+        $(SRC_DIR)/zappy_net_io.c \
+        $(SRC_DIR)/zappy_net_readline.c
 SRC_CPP =
 
 SRC_FILES = $(SRC_C) $(SRC_CPP)
